move all tests to ___test___  folder
