+++ src/__tests__/CurrentQuotes.test.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import Quote from './quote';
+import CurrentQuote from '../components/CurrentQuotes';
 
 it('renders without crashing', () => {
   const div = document.createElement('div');
-  ReactDOM.render(<Quote />, div);
+  ReactDOM.render(<CurrentQuote />, div);
 });
