+++ src/__tests__/home.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
-import Home from './home';
+import Home from '../pages/home';
 
 test('match home snapshot', () => {
   const tree = renderer.create(<Home />).toJSON();
