+++ src/__tests__/operate.test.js
@@ -1,4 +1,4 @@
-import operate from './operate';
+import operate from '../logic/operate';
 
 describe('operate', () => {
   test('shoud return 10 after adding 7 + 3', () => {
