+++ src/__tests__/quote.test.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import CurrentQuote from './CurrentQuotes';
+import Quote from '../pages/quote';
 
 it('renders without crashing', () => {
   const div = document.createElement('div');
-  ReactDOM.render(<CurrentQuote />, div);
+  ReactDOM.render(<Quote />, div);
 });
